fix : 가게 CRUD 기능 수정

+++ src/main/java/com/team2/grabtable/domain/store/dto/StoreDto.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @Builder
 public class StoreDto {
-    private int storeId;
-    private int ownerId;
+    private Long storeId;
+    private Long ownerId;
     private String name;
     private String location;
     private String type;
